chore: move more ember specific implementaitons to `migration-graph-ember`
- Move `ProjectGraph` to `migration-graph-shared`.
- Create `EmberAppPackageGraph`, `EmberAddonPackageGraph`
- Create `EmberAppProjectGraph`, `EmberAddonProjectGraph`
