+++ packages/migration-graph-ember/package.json
@@ -24,6 +24,7 @@
     "@rehearsal/migration-graph-shared": "^0.0.37",
     "@swc/core": "^1.3.9",
     "debug": "^4.3.2",
+    "enhanced-resolve": "^5.10.0",
     "fast-glob": "^3.2.12",
     "find-up": "^5.0.0",
     "fs-extra": "^10.1.0",
@@ -34,7 +35,13 @@
   },
   "devDependencies": {
     "@babel/types": "^7.18.13",
+    "@rehearsal/test-support": "workspace:*",
+    "@types/lodash.merge": "^4.6.7",
+    "@types/rimraf": "^3.0.2",
+    "dependency-cruiser": "^11.15.0",
     "fixturify": "^2",
+    "lodash.merge": "^4.6.2",
+    "rimraf": "^3.0.2",
     "tmp": "^0.2.1",
     "vitest": "^0.23.4",
     "walk-sync": "^3"
