+++ packages/migration-graph-ember/src/entities/ember-addon-project-graph.ts
@@ -1,0 +1,10 @@
+import debug from 'debug';
+import { EmberAppProjectGraph } from './ember-app-project-graph';
+const DEBUG_CALLBACK = debug('rehearsal:ember-addon-project-graph');
+
+export class EmberAddonProjectGraph extends EmberAppProjectGraph {
+  constructor(rootDir: string, sourceType: string) {
+    super(rootDir, sourceType);
+    DEBUG_CALLBACK('constructor');
+  }
+}
