+++ packages/migration-graph-ember/src/index.ts
@@ -21,3 +21,15 @@ export {
 
 export { discoverEmberPackages } from './utils/discover-ember-packages';
 export { discoverServiceDependencies } from './utils/discover-ember-service-dependencies';
+
+export {
+  EmberAppPackageGraph,
+  EmberAppPackageyGraphOptions,
+} from './entities/ember-app-package-graph';
+export {
+  EmberAddonPackageGraph,
+  EmberAddonPackageGraphOptions,
+} from './entities/ember-addon-package-graph';
+
+export { EmberAppProjectGraph } from './entities/ember-app-project-graph';
+export { EmberAddonProjectGraph } from './entities/ember-addon-project-graph';
