+++ packages/migration-graph-shared/src/graph/index.ts
@@ -1,0 +1,3 @@
+export { dfs } from './dfs';
+export { Graph } from './graph';
+export { GraphNode } from './node';
