+++ packages/migration-graph-shared/src/index.ts
@@ -1,9 +1,7 @@
 import { resolve } from 'path';
 import { readJsonSync } from 'fs-extra';
 
-import { PackageJson } from './entities/package';
-
-export function readPackageJson(pathToPackage: string): PackageJson {
+export function readPackageJson(pathToPackage: string): Record<string, any> {
   return readJsonSync(resolve(pathToPackage, 'package.json'));
 }
 
@@ -17,3 +15,8 @@ export { formatter } from './utils/prettier';
 export { isWorkspace } from './utils/workspace';
 export { setNestedPropertyValue, removeNestedPropertyValue } from './utils/pojo';
 export { isTesting, setupTestEnvironment } from './utils/environment';
+
+export * from './graph';
+export * from './types';
+export { PackageGraph, PackageGraphOptions } from './entities/package-graph';
+export { ProjectGraph } from './entities/project-graph';
