+++ packages/migration-graph/src/index.ts
@@ -1,7 +1,12 @@
 export { buildMigrationGraph } from './migration-graph';
 export { discoverEmberPackages } from '@rehearsal/migration-graph-ember';
-export { GraphNode } from './utils/graph-node';
-export { Graph } from './utils/graph';
+export type {
+  UniqueNode,
+  PackageNode,
+  ModuleNode,
+  GraphNode,
+} from '@rehearsal/migration-graph-shared';
+export { Graph } from '@rehearsal/migration-graph-shared';
 
 export {
   getMigrationStrategy,
@@ -9,5 +14,3 @@ export {
   type MigrationStrategyOptions,
   type SourceFile,
 } from './migration-strategy';
-
-export type { UniqueGraphNode, PackageNode, ModuleNode } from './types';
