+++ packages/migration-graph/src/source-type.ts
@@ -1,0 +1,5 @@
+export enum SourceType {
+  EmberApp = 'Ember Application',
+  EmberAddon = 'Ember Addon',
+  Library = 'Javascript Library',
+}
